More indent-related fixes

+++ yt/python/yson/yson.cpp
@@ -518,7 +518,7 @@ private:
         const int maxIndentValue = 128;
 
         if (HasArgument(args, kwargs, "indent")) {
-            auto arg = Py::Int(ExtractArgument(args, kwargs, "indent"));
+            auto arg = Py::Long(ExtractArgument(args, kwargs, "indent"));
             auto longIndent = arg.as_long();
             if (longIndent > maxIndentValue) {
                 throw CreateYsonError(Format("Indent value exceeds indentation limit: %v > %v",
